SCRUM-4 : client management done

+++ src/main/java/com/wora/psql.sql
@@ -12,15 +12,15 @@ CREATE TABLE clients(
 CREATE TABLE projects(
     id uuid PRIMARY KEY,
     project_name VARCHAR(255),
-    profit_margin DOUBLE,
-    total_cost DOUBLE,
+    profit_margin FLOAT,
+    total_cost FLOAT,
     project_status project_status,
     client_id uuid REFERENCES clients(id)
 );
 
 CREATE TABLE estimates(
     id uuid PRIMARY KEY,
-    estimated_amount DOUBLE,
+    estimated_amount FLOAT,
     issue_date DATE,
     validity_date DATE,
     is_accepted BOOLEAN,
@@ -29,22 +29,22 @@ CREATE TABLE estimates(
 
 CREATE TABLE components(
     id uuid PRIMARY KEY,
-    tva DOUBLE,
+    tva FLOAT,
     component_type component_type,
     project_id uuid REFERENCES projects(id)
 );
 
 CREATE table workers(
     id uuid PRIMARY KEY,
-    hourly_rate DOUBLE,
-    work_hours DOUBLE,
-    worker_productivity DOUBLE
+    hourly_rate FLOAT,
+    work_hours FLOAT,
+    worker_productivity FLOAT
 ) INHERITS(components);
 
 CREATE TABLE materials(
     id uuid PRIMARY KEY,
-    unit_cost DOUBLE,
-    quantity DOUBLE,
-    transport_cost DOUBLE,
-    quality_coefficient DOUBLE
+    unit_cost FLOAT,
+    quantity FLOAT,
+    transport_cost FLOAT,
+    quality_coefficient FLOAT
 ) INHERITS(components);
